requirements.txt
absl-py==2.1.0
altair==5.4.1
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dbfread==2.0.7
flatbuffers==24.3.25
folium==0.18.0
fonttools==4.54.1
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
narwhals==1.11.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.3
psycopg2==2.9.10
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
referencing==0.35.1
reportlab==4.4.0
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
simpledbf==0.2.6
six==1.16.0
smmap==5.0.1
streamlit==1.39.0
streamlit_folium==0.23.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
Werkzeug==3.1.1
wheel==0.44.0
wrapt==1.16.0
XlsxWriter==3.2.3
xyzservices==2024.9.0
